% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridwell.R
\docType{package}
\name{gridwell-package}
\alias{gridwell}
\alias{gridwell-package}
\title{gridwell: Fast Multi-Format Table Rendering}
\description{
Renders table IR (Intermediate Representation) to many output formats including HTML, LaTeX, Typst, RTF, SVG, ANSI, Pandoc, Quarto, DOCX, XLSX, and PPTX. Powered by a Rust backend via extendr.
}
\author{
\strong{Maintainer}: Richard Iannone \email{riannone@me.com}

}
\keyword{internal}
