% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_emitter.R
\name{gt_to_ir}
\alias{gt_to_ir}
\title{Convert a gt table object to gridwell IR JSON}
\usage{
gt_to_ir(gt_obj)
}
\arguments{
\item{gt_obj}{A \code{gt_tbl} object from the gt package.}
}
\value{
A JSON string in gridwell IR format.
}
\description{
Convert a gt table object to gridwell IR JSON
}
