% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_emitter.R
\name{gt_to_ir_list}
\alias{gt_to_ir_list}
\title{Convert a gt table object to a gridwell IR list}
\usage{
gt_to_ir_list(gt_obj)
}
\arguments{
\item{gt_obj}{A \code{gt_tbl} object from the gt package.}
}
\value{
A list representing the gridwell IR.
}
\description{
Convert a gt table object to a gridwell IR list
}
