% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/gridwell.R
\name{gw_render_ansi}
\alias{gw_render_ansi}
\title{Render a table with ANSI escape codes.}
\usage{
gw_render_ansi(table_ptr)

gw_render_ansi(table_ptr)
}
\arguments{
\item{table_ptr}{An external pointer to a parsed table.}
}
\value{
A string with ANSI codes.

A string with ANSI codes.
}
\description{
Render a table with ANSI escape codes.

Render a table with ANSI escape codes.
}
