% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridwell.R
\name{gw_render_format}
\alias{gw_render_format}
\alias{gw_render_html}
\alias{gw_render_latex}
\alias{gw_render_typst}
\alias{gw_render_rtf}
\alias{gw_render_svg}
\alias{gw_render_ansi}
\alias{gw_render_pandoc}
\alias{gw_render_quarto}
\alias{gw_render_docx}
\alias{gw_render_xlsx}
\alias{gw_render_pptx}
\title{Render a table to a specific format.}
\usage{
gw_render_html(table_ptr, ...)

gw_render_latex(table_ptr, ...)

gw_render_typst(table_ptr, ...)

gw_render_rtf(table_ptr, ...)

gw_render_svg(table_ptr, ...)

gw_render_ansi(table_ptr, ...)

gw_render_pandoc(table_ptr, ...)

gw_render_quarto(table_ptr, ...)

gw_render_docx(table_ptr, ...)

gw_render_xlsx(table_ptr, ...)

gw_render_pptx(table_ptr, ...)
}
\arguments{
\item{table_ptr}{An external pointer to a parsed table.}

\item{...}{Writer options as named arguments (see \code{\link[=gw_formats]{gw_formats()}}).}
}
\value{
A string for text formats, a raw vector for binary formats.
}
\description{
Shorthands for \code{\link[=gw_render]{gw_render()}} and \code{\link[=gw_render_binary]{gw_render_binary()}}; writer options are
passed as named arguments, e.g. \code{gw_render_html(tbl, inline_styles = TRUE)}.
}
