% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/gridwell.R
\name{gw_render_html}
\alias{gw_render_html}
\title{Render a table to HTML.}
\usage{
gw_render_html(table_ptr)

gw_render_html(table_ptr)
}
\arguments{
\item{table_ptr}{An external pointer to a parsed table.}
}
\value{
An HTML string.

An HTML string.
}
\description{
Render a table to HTML.

Render a table to HTML.
}
