% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/gridwell.R
\name{gw_render_latex}
\alias{gw_render_latex}
\title{Render a table to LaTeX.}
\usage{
gw_render_latex(table_ptr)

gw_render_latex(table_ptr)
}
\arguments{
\item{table_ptr}{An external pointer to a parsed table.}
}
\value{
A LaTeX string.

A LaTeX string.
}
\description{
Render a table to LaTeX.

Render a table to LaTeX.
}
