% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/gridwell.R
\name{gw_render_pandoc}
\alias{gw_render_pandoc}
\title{Render a table to Pandoc AST JSON.}
\usage{
gw_render_pandoc(table_ptr)

gw_render_pandoc(table_ptr)
}
\arguments{
\item{table_ptr}{An external pointer to a parsed table.}
}
\value{
A Pandoc JSON string.

A Pandoc JSON string.
}
\description{
Render a table to Pandoc AST JSON.

Render a table to Pandoc AST JSON.
}
