% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/gridwell.R
\name{gw_render_quarto}
\alias{gw_render_quarto}
\title{Render a table to Quarto-flavored Markdown.}
\usage{
gw_render_quarto(table_ptr)

gw_render_quarto(table_ptr)
}
\arguments{
\item{table_ptr}{An external pointer to a parsed table.}
}
\value{
A Quarto Markdown string.

A Quarto Markdown string.
}
\description{
Render a table to Quarto-flavored Markdown.

Render a table to Quarto-flavored Markdown.
}
