% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/gridwell.R
\name{gw_render_rtf}
\alias{gw_render_rtf}
\title{Render a table to RTF.}
\usage{
gw_render_rtf(table_ptr)

gw_render_rtf(table_ptr)
}
\arguments{
\item{table_ptr}{An external pointer to a parsed table.}
}
\value{
An RTF string.

An RTF string.
}
\description{
Render a table to RTF.

Render a table to RTF.
}
