% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/gridwell.R
\name{gw_render_svg}
\alias{gw_render_svg}
\title{Render a table to SVG.}
\usage{
gw_render_svg(table_ptr)

gw_render_svg(table_ptr)
}
\arguments{
\item{table_ptr}{An external pointer to a parsed table.}
}
\value{
An SVG string.

An SVG string.
}
\description{
Render a table to SVG.

Render a table to SVG.
}
