% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/gridwell.R
\name{gw_render_typst}
\alias{gw_render_typst}
\title{Render a table to Typst.}
\usage{
gw_render_typst(table_ptr)

gw_render_typst(table_ptr)
}
\arguments{
\item{table_ptr}{An external pointer to a parsed table.}
}
\value{
A Typst string.

A Typst string.
}
\description{
Render a table to Typst.

Render a table to Typst.
}
