% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridwell.R
\name{gw_to_json}
\alias{gw_to_json}
\title{Serialize a parsed table back to JSON.}
\usage{
gw_to_json(table_ptr)
}
\arguments{
\item{table_ptr}{An external pointer to a parsed table.}
}
\value{
A JSON string.
}
\description{
Serialize a parsed table back to JSON.
}
