% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridwell.R
\name{gw_validate}
\alias{gw_validate}
\title{Validate a parsed table.}
\usage{
gw_validate(table_ptr)
}
\arguments{
\item{table_ptr}{An external pointer to a parsed table (from \code{gw_parse_ir()}).}
}
\value{
A character vector of validation errors (empty if valid).
}
\description{
Validate a parsed table.
}
